==> notebook/05_personal_recommendation_test.ipynb <==
# Personal Movies Recommendation

## Imports

import pandas as pd
from src.svd import run_svd_pipeline

## Dataframe Creation

ratings = pd.read_csv("../data/processed/ratings.csv")
movies = pd.read_csv("../data/processed/movies.csv")

my_movies = pd.DataFrame({
    'item_id': [127, 182, 56, 178],
    'my_rating': [4, 5, 3, 3]
})

my_movies = my_movies.merge(movies[['item_id', 'title']], on='item_id')
print("You've Already Watched:\n")
display(my_movies[['title', 'my_rating']])


## SVD Recommendation

# 4. Run SVD
user_id = 1  # Could be any dummy ID
top_k, _, _, _ = run_svd_pipeline(user_id=user_id, ratings_df=ratings, k=30)

# 5. Build recommendation DataFrame
recs_df = pd.DataFrame(top_k, columns=["item_id", "predicted_rating"])

# 6. Merge with movie titles
recs_df = recs_df.merge(movies[['item_id', 'title']], on="item_id")

# 7. Remove already-watched items
filtered_recs = recs_df[~recs_df['item_id'].isin(my_movies['item_id'])]

# 8. Show final top-N recommendations
top_n = 10
print(f"\nTop {top_n} Recommended Movies:\n")
display(filtered_recs[['title', 'predicted_rating']].head(top_n))

## Item-Based Collaborative Filtering Recommendation

from src.item_cf import run_item_cf_pipeline

# Run Item-Based CF
item_cf_top_k, _ = run_item_cf_pipeline(user_id=user_id, ratings_df=ratings, k=30)

# Convert to DataFrame
item_cf_df = pd.DataFrame(item_cf_top_k, columns=["item_id", "predicted_score"])

# Merge with movie titles
item_cf_df = item_cf_df.merge(movies[['item_id', 'title']], on='item_id')

# Filter out already-watched movies
item_cf_filtered = item_cf_df[~item_cf_df['item_id'].isin(my_movies['item_id'])]

# Show top-N item-based recommendations
print(f"\nTop {top_n} Item-Based CF Recommendations:\n")
display(item_cf_filtered[['title', 'predicted_score']].head(top_n))

## User-Based Collaborative Filtering Recommendation

from src.user_cf import run_user_cf_pipeline

# Run User-Based CF
user_cf_top_k, _, _ = run_user_cf_pipeline(user_id=user_id, ratings_df=ratings, k=30)

# Convert to DataFrame
user_cf_df = pd.DataFrame(user_cf_top_k, columns=["item_id", "predicted_rating"])

# Merge with movie titles
user_cf_df = user_cf_df.merge(movies[['item_id', 'title']], on="item_id")

# Filter out already-watched movies
user_cf_filtered = user_cf_df[~user_cf_df['item_id'].isin(my_movies['item_id'])]

# Show top-N user-based recommendations
print(f"\nTop {top_n} User-Based CF Recommendations:\n")
display(user_cf_filtered[['title', 'predicted_rating']].head(top_n))
